+++ Cuda/Src/intList.cu
@@ -69,7 +69,7 @@ void printIntList(const Int_List list) {
 }
 */
 
-__device__ void copyTabDev(int *src, int *dest, int size) {
+__device__ void copyTabDev(uint64_t *src, uint64_t *dest, int size) {
 	if (blockIdx.x == 0) {
 		int tid = threadIdx.x;
 		if (tid < size) {
@@ -79,7 +79,7 @@ __device__ void copyTabDev(int *src, int *dest, int size) {
 }
 
 //1 block ; size thread
-__global__ void copyTabGPU(int *src, int *dest, int size) {
+__global__ void copyTabGPU(uint64_t *src, uint64_t *dest, int size) {
 	int tid = threadIdx.x;
 	if (tid < size) {
 		dest[tid] = src[tid];
@@ -95,14 +95,14 @@ __host__ Int_List_GPU *createIntList() {
 
 __host__ void addInt(Int_List_GPU **list, int v) {
 	Int_List_GPU *l = new Int_List_GPU[1];
-	l->List = new int[(*list)->Size + 1];
+	l->List = new uint64_t[(*list)->Size + 1];
 	l->Size = (*list)->Size + 1;
-	int *dev_list_dest, *dev_list_src;
-	cudaMalloc((void **)&dev_list_src, (*list)->Size * sizeof(int));
-	cudaMemcpy(dev_list_src, (*list)->List, (*list)->Size * sizeof(int), cudaMemcpyHostToDevice);
-	cudaMalloc((void **)&dev_list_dest, ((*list)->Size + 1) * sizeof(int));
+	uint64_t *dev_list_dest, *dev_list_src;
+	cudaMalloc((void **)&dev_list_src, (*list)->Size * sizeof(uint64_t));
+	cudaMemcpy(dev_list_src, (*list)->List, (*list)->Size * sizeof(uint64_t), cudaMemcpyHostToDevice);
+	cudaMalloc((void **)&dev_list_dest, ((*list)->Size + 1) * sizeof(uint64_t));
 	copyTabGPU<<<1, (*list)->Size>>>(dev_list_src, dev_list_dest, (*list)->Size);
-	cudaMemcpy(l->List, dev_list_dest, (*list)->Size * sizeof(int), cudaMemcpyDeviceToHost);
+	cudaMemcpy(l->List, dev_list_dest, (*list)->Size * sizeof(uint64_t), cudaMemcpyDeviceToHost);
 	cudaFree(dev_list_src);
 	cudaFree(dev_list_dest);
 	l->List[(*list)->Size] = v;
@@ -112,11 +112,11 @@ __host__ void addInt(Int_List_GPU **list, int v) {
 	*list = l;
 }
 
-__device__ void addIntGPU(int **list, int size, int v) {
+__device__ void addIntGPU(uint64_t **list, int size, int v) {
 	if (blockIdx.x == 0) {
-		__shared__ int *l;
+		__shared__ uint64_t *l;
 		if (threadIdx.x == 0) {
-			l = new int[size + 1];
+			l = new uint64_t[size + 1];
 		}
 		copyTabDev((*list), l, size);
 		l[size] = v;
@@ -125,17 +125,17 @@ __device__ void addIntGPU(int **list, int size, int v) {
 	}
 }
 
-__host__ int getVal(Int_List_GPU l, int index) {
+__host__ uint64_t getVal(Int_List_GPU l, int index) {
 	return l.List[index];
 }
 
-__device__ int getValGPU(int *l, int index) {
+__device__ uint64_t getValGPU(uint64_t *l, int index) {
 	return l[index];
 }
 
 __host__ void removeLastInt(Int_List_GPU **list) {
 	Int_List_GPU *l = new Int_List_GPU[1];
-	l->List = new int[(*list)->Size - 1];
+	l->List = new uint64_t[(*list)->Size - 1];
 	int *dev_list_dest, *dev_list_src;
 	cudaMalloc((void **)&dev_list_src, (*list)->Size * sizeof(int));
 	cudaMemcpy(dev_list_src, (*list)->List, (*list)->Size * sizeof(int), cudaMemcpyHostToDevice);
@@ -150,11 +150,11 @@ __host__ void removeLastInt(Int_List_GPU **list) {
 	*list = l;
 }
 
-__device__ void removeLastInt(int **list, int size) {
+__device__ void removeLastInt(uint64_t **list, uint64_t size) {
 	if (blockIdx.x == 0) {
-		__shared__ int *l;
+		__shared__ uint64_t *l;
 		if (threadIdx.x == 0) {
-			l = new int[size - 1];
+			l = new uint64_t[size - 1];
 		}
 		copyTabDev((*list), l, size - 1);
 		//delete(list);
@@ -168,9 +168,9 @@ __host__ void resetIntList(Int_List_GPU **list) {
 	}
 }
 
-__device__ void resetIntListGPU(int **list, int size) {
+__device__ void resetIntListGPU(uint64_t **list, uint64_t size) {
 	if (blockIdx.x == 0) {
-		for (int i = 0; i < size; i++) {
+		for (uint64_t i = 0; i < size; i++) {
 			removeLastInt(list, size);
 		}
 	}
@@ -178,7 +178,7 @@ __device__ void resetIntListGPU(int **list, int size) {
 
 __host__ void printIntList(Int_List_GPU l) {
 	printf("Taille de la liste : %i\n", l.Size);
-	for (int i = 0; i < l.Size; i++) {
-		printf("valeur : %i\n", getVal(l, i));
+	for (uint64_t i = 0; i < l.Size; i++) {
+		printf("valeur : %ld\n", getVal(l, i));
 	}
 }
